fix conflicts with AzkabanSpecExecutor
